# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amenly",
    "description": "Amenly AI GRC Platform - Production API",
    "version": "1.0.1"
  },
  "paths": {
...
    "/api/v1/rag/query": {
      "post": {
        "tags": [
          "RAG System",
          "RAG System"
        ],
        "summary": "Query Rag",
        "description": "Query RAG system with compliance question\n\nThis endpoint:\n1. Generates query embedding\n2. Searches Qdrant for relevant chunks\n3. Builds context from retrieved documents\n4. Generates answer using LLM\n5. Returns structured response with sources\n\nExample:\n    ```json\n    {\n        \"question\": \"What does ISO 27001 require for MFA?\",\n        \"framework\": \"ISO27001\",\n        \"top_k\": 5\n    }\n    ```",
        "operationId": "query_rag_api_v1_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGQueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      }
    },
    "/api/v1/rag/search": {
      "post": {
        "tags": [
          "RAG System",
          "RAG System"
        ],
        "summary": "Search Documents",
        "description": "Semantic search without LLM generation\n\nReturns relevant document chunks without generating an answer.\nUseful for:\n- Exploring available content\n- Finding specific controls or sections\n- Debugging retrieval quality",
        "operationId": "search_documents_api_v1_rag_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RAGSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAGSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "login": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RAGQueryRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Question",
            "description": "User question"
          },
          "framework": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__ai__rag__schemas__FrameworkType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Filter by specific framework",
            "default": "ALL"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "Number of chunks to retrieve",
            "default": 5
          },
          "score_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score Threshold",
            "description": "Minimum similarity score",
            "default": 0.5
          },
          "include_metadata": {
            "type": "boolean",
            "title": "Include Metadata",
            "description": "Include chunk metadata in response",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "RAGQueryRequest",
        "description": "Request for RAG query",
        "example": {
          "framework": "ISO27001",
          "include_metadata": true,
          "question": "What does ISO 27001 require for MFA?",
          "score_threshold": 0.5,
          "top_k": 5
        }
      },
      "RAGQueryResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Query success status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response data"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "RAGQueryResponse",
        "description": "Response for RAG query",
        "example": {
          "data": {
            "answer": "ISO 27001 requires multi-factor authentication...",
            "confidence_score": 0.89,
            "processing_time_ms": 1250,
            "retrieved_chunks": 5,
            "sources": [
              {
                "control_id": "A.9.4.2",
                "framework": "ISO27001",
                "page_number": 45,
                "relevance_score": 0.92,
                "section": "A.9.4.2",
                "source_file": "iso27001_2022.pdf"
              }
            ]
          },
          "message": "Query processed successfully",
          "success": true
        }
      },
      "RAGSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Query",
            "description": "Search query"
          },
          "framework": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__ai__rag__schemas__FrameworkType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Filter by framework",
            "default": "ALL"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "Number of results",
            "default": 10
          },
          "score_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score Threshold",
            "description": "Minimum score",
            "default": 0.3
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RAGSearchRequest",
        "description": "Request for semantic search only (no LLM generation)",
        "example": {
          "framework": "ISO27001",
          "query": "access control requirements",
          "score_threshold": 0.3,
          "top_k": 10
        }
      },
      "RAGSearchResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "RAGSearchResponse",
        "description": "Response for semantic search",
        "example": {
          "data": {
            "query_time_ms": 45,
            "results": [
              {
                "metadata": {
                  "framework": "ISO27001",
                  "section": "A.9"
                },
                "score": 0.92,
                "text": "Access control requirements..."
              }
            ],
            "total_results": 10
          },
          "message": "Search completed successfully",
          "success": true
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__ai__rag__schemas__FrameworkType": {
        "type": "string",
        "enum": [
          "ISO27001",
          "NIST",
          "SOC2",
          "GDPR",
          "HIPAA",
          "PCI_DSS",
          "COBIT",
          "TISAX",
          "ALL"
        ],
        "title": "FrameworkType",
        "description": "Supported compliance frameworks"
      },
...
    },
    "securitySchemes": {
      "login": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}